Conflicts:
	src/main/java/com/greendelta/collaboration/util/MetaData.java
	src/main/java/com/greendelta/collaboration/util/ModelTypes.java

+++ src/main/java/com/greendelta/collaboration/service/LibraryService.java
@@ -7,6 +7,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -21,12 +22,14 @@ import org.openlca.util.Dirs;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.greendelta.collaboration.error.ForbiddenAccessException;
 import com.greendelta.collaboration.error.ServiceUnavailableException;
 import com.greendelta.collaboration.model.LibraryAccess;
 import com.greendelta.collaboration.model.User;
 import com.greendelta.collaboration.model.settings.LibrarySetting;
 import com.greendelta.collaboration.model.settings.ServerSetting;
 import com.greendelta.collaboration.model.settings.SettingType;
+import com.greendelta.collaboration.service.user.TeamService;
 import com.greendelta.collaboration.service.user.UserService;
 
 @Service
@@ -35,36 +38,54 @@ public class LibraryService {
 	private static final Logger log = LogManager.getLogger(LibraryService.class);
 	private final UserService userService;
 	private final RepositoryService repoService;
+	private final TeamService teamService;
 	private final SettingsService settings;
 
 	@Autowired
-	public LibraryService(UserService userService, RepositoryService repoService, SettingsService settings) {
+	public LibraryService(UserService userService, RepositoryService repoService, TeamService teamService,
+			SettingsService settings) {
 		this.userService = userService;
 		this.repoService = repoService;
+		this.teamService = teamService;
 		this.settings = settings;
 	}
 
 	public List<String> getAllAccessible() {
-		String libraryPath = settings.get(ServerSetting.LIBRARY_PATH);
+		return getAccessibleForTeams(new ArrayList<>());
+	}
+
+	public List<String> getAccessibleForTeams() {
+		var currentUser = userService.getCurrentUser();
+		var teams = teamService.getTeamsFor(currentUser);
+		if (teams.isEmpty())
+			return new ArrayList<>();
+		return getAccessibleForTeams(teams.stream().map(team -> team.teamname).toList());
+	}
+
+	public List<String> getAccessibleForTeams(List<String> teams) {
+		var libraryPath = getLibraryPath();
 		if (libraryPath == null)
 			return new ArrayList<String>();
 		var accessCheck = new AccessCheck();
 		var libraries = new ArrayList<String>();
 		for (var file : new File(libraryPath).listFiles()) {
-			if (!file.getName().endsWith(".zip"))
+			var name = file.getName();
+			if (!name.endsWith(".zip"))
 				continue;
-			var id = file.getName().substring(0, file.getName().length() - 4);
-			if (!accessCheck.canAccess(id))
+			var id = name.substring(0, name.lastIndexOf(".zip"));
+			if (!accessCheck.canAccess(id, teams))
 				continue;
 			libraries.add(id);
 		}
 		return libraries;
 	}
 
-	public String insert(InputStream stream, LibraryAccess access) throws IOException {
+	public String insert(InputStream stream, String access) throws IOException {
 		if (stream == null)
 			return null;
-		var libraryPath = getLibraryPath();
+		var currentUser = userService.getCurrentUser();
+		if (!currentUser.isDataManager() && (!LibraryAccess.isTeamAccess(access) || !isTeamMember(access, currentUser)))
+			throw new ForbiddenAccessException("LIBRARIES", "WRITE");
 		Path tmpFile = null;
 		try {
 			tmpFile = Files.createTempFile("cs-lib-", ".zip");
@@ -72,12 +93,14 @@ public class LibraryService {
 			var info = LibraryPackage.getInfo(tmpFile.toFile());
 			if (info == null)
 				return null;
-			var file = new File(libraryPath, info.name() + ".zip");
-			if (file.exists())
+			var id = info.name();
+			var file = getLibraryFile(id);
+			if (file.exists() && getAccessTypes(id).contains(access))
 				throw new IOException("existed");
-			Files.copy(tmpFile, file.toPath());
-			settings.get(SettingType.LIBRARY_SETTING, info.name(), settings.ACCESS.DATA_MANAGER)
-					.set(LibrarySetting.ACCESS, access);
+			if (!file.exists()) {
+				Files.copy(tmpFile, file.toPath());
+			}
+			addAccessType(info.name(), access);
 			return info.name();
 		} catch (IOException e) {
 			if (!"existed".equals(e.getMessage())) {
@@ -89,17 +112,16 @@ public class LibraryService {
 		}
 	}
 
-	public void update(String id, LibraryAccess access) {
-		settings.get(SettingType.LIBRARY_SETTING, id, settings.ACCESS.DATA_MANAGER)
-				.set(LibrarySetting.ACCESS, access);
-	}
-
-	public boolean delete(String id) {
-		var file = new File(getLibraryPath(), id + ".zip");
+	public boolean delete(String id, String access) {
+		checkWriteAccess(id, Arrays.asList(access));
+		var file = getLibraryFile(id);
 		if (!file.exists())
 			return false;
 		try {
-			Files.delete(file.toPath());
+			var accessTypes = removeAccessType(id, access);
+			if (accessTypes.isEmpty()) {
+				Files.delete(file.toPath());
+			}
 			return true;
 		} catch (IOException e) {
 			log.error("Error deleting library from disk", e);
@@ -107,15 +129,38 @@ public class LibraryService {
 		}
 	}
 
+	private void checkWriteAccess(String id, List<String> accessTypes) {
+		var currentUser = userService.getCurrentUser();
+		if (currentUser.isDataManager())
+			return;
+		for (var access : accessTypes) {
+			if (LibraryAccess.isTeamAccess(access)) {
+				var team = teamService.getForTeamname(access);
+				if (team != null && team.users.contains(currentUser))
+					return;
+			}
+		}
+		throw new ForbiddenAccessException(id, "WRITE");
+	}
+
+	private boolean isTeamMember(String teamname, User user) {
+		var team = teamService.getForTeamname(teamname);
+		if (team == null)
+			return false;
+		return team.users.contains(user);
+	}
+
 	public File get(String id) {
-		var file = new File(getLibraryPath(), id + ".zip");
+		if (!new AccessCheck().canAccess(id, new ArrayList<>()))
+			throw new ForbiddenAccessException(id, "READ");
+		var file = getLibraryFile(id);
 		if (!file.exists())
 			return null;
 		return file;
 	}
 
-	public LibraryInfo getInfo(String id) {
-		var file = new File(getLibraryPath(), id + ".zip");
+	public LibraryInfo getInfo(String id, boolean isAdminArea) {
+		var file = getLibraryFile(id);
 		if (!file.exists())
 			return null;
 		try (var repos = repoService.getAllAccessible()) {
@@ -123,8 +168,20 @@ public class LibraryService {
 					.filter(repo -> repo.linkedLibraries().contains(id))
 					.map(Repository::toId)
 					.distinct().toList();
-			return new LibraryInfo(LibraryPackage.getInfo(file), linkedIn);
+			var accesses = getAccessTypes(id);
+			if (!isAdminArea) {
+				for (var access : LibraryAccess.values()) {
+					accesses.remove(access.name());
+				}
+			}
+			if (accesses.isEmpty())
+				return null;
+			return new LibraryInfo(LibraryPackage.getInfo(file), linkedIn, accesses);
 		}
+	}
+
+	private File getLibraryFile(String id) {
+		return new File(getLibraryPath(), id + ".zip");
 	}
 
 	private String getLibraryPath() {
@@ -132,6 +189,42 @@ public class LibraryService {
 		if (libraryPath == null)
 			throw new ServiceUnavailableException("Library service unavailable because library path is not set");
 		return libraryPath;
+	}
+
+	private List<String> getAccessTypes(String id) {
+		return settings.get(SettingType.LIBRARY_SETTING, id, settings.ACCESS.USER)
+				.get(LibrarySetting.ACCESS, new ArrayList<>());
+	}
+
+	private List<String> addAccessType(String id, String access) {
+		var settingsAccess = LibraryAccess.isTeamAccess(access)
+				? settings.ACCESS.TEAM_DATA(teamService.getForTeamname(access))
+				: settings.ACCESS.DATA_MANAGER;
+		var accessTypes = getAccessTypes(id);
+		if (accessTypes.contains(access))
+			return accessTypes;
+		accessTypes.add(access);
+		settings.get(SettingType.LIBRARY_SETTING, id, settingsAccess)
+				.set(LibrarySetting.ACCESS, accessTypes);
+		return accessTypes;
+	}
+
+	private List<String> removeAccessType(String id, String access) {
+		var settingsAccess = LibraryAccess.isTeamAccess(access)
+				? settings.ACCESS.TEAM_DATA(teamService.getForTeamname(access))
+				: settings.ACCESS.DATA_MANAGER;
+		var accessTypes = getAccessTypes(id);
+		if (!accessTypes.contains(access))
+			return accessTypes;
+		accessTypes.remove(access);
+		if (accessTypes.isEmpty()) {
+			settings.get(SettingType.LIBRARY_SETTING, id, settingsAccess)
+			.set(LibrarySetting.ACCESS, null);
+		} else {
+			settings.get(SettingType.LIBRARY_SETTING, id, settingsAccess)
+					.set(LibrarySetting.ACCESS, accessTypes);
+		}
+		return accessTypes;
 	}
 
 	private class AccessCheck {
@@ -143,41 +236,54 @@ public class LibraryService {
 			this.user = userService.getCurrentUser();
 		}
 
-		private boolean canAccess(String library) {
+		private boolean canAccess(String library, List<String> teamnames) {
+			var accesses = getAccessTypes(library);
+			if (accesses.isEmpty())
+				return false;
+			if (!teamnames.isEmpty())
+				return teamnames.stream()
+						.filter(accesses::contains)
+						.filter(access -> isTeamMember(access, user))
+						.count() > 0;
 			if (user.isDataManager())
 				return true;
-			LibraryAccess access = settings
-					.get(SettingType.LIBRARY_SETTING, library, settings.ACCESS.DATA_MANAGER)
-					.get(LibrarySetting.ACCESS);
-			if (access == null)
-				return false;
-			return switch (access) {
-				case PUBLIC -> true;
-				case USER -> !user.isAnonymous();
-				case MEMBER -> linkedLibraries().contains(library);
-			};
+			for (var access : accesses) {
+				var hasAccess = switch (access) {
+					case "PUBLIC" -> true;
+					case "USER" -> !user.isAnonymous();
+					case "MEMBER" -> linkedLibraries().contains(library);
+					default -> isTeamMember(access, user);
+				};
+				if (hasAccess)
+					return true;
+			}
+			return false;
 		}
 
 		private Set<String> linkedLibraries() {
 			if (linkedLibraries != null)
 				return linkedLibraries;
-			linkedLibraries = repoService.getAllAccessible().stream()
-					.map(Repository::gitRepo)
-					.map(Repositories::infoOf)
-					.filter(Objects::nonNull)
-					.map(PackageInfo::libraries)
-					.flatMap(List::stream)
-					.distinct()
-					.collect(Collectors.toSet());
+			try (var accessible = repoService.getAllAccessible()) {
+				linkedLibraries = accessible.stream()
+						.map(Repository::gitRepo)
+						.map(Repositories::infoOf)
+						.filter(Objects::nonNull)
+						.map(PackageInfo::libraries)
+						.flatMap(List::stream)
+						.distinct()
+						.collect(Collectors.toSet());
+			}
 			return linkedLibraries;
 		}
 
 	}
 
-	public record LibraryInfo(String name, String description, boolean isRegionalized, List<String> linkedIn) {
+	public record LibraryInfo(String name, String description, boolean isRegionalized, List<String> linkedIn,
+			List<String> accessTypes) {
 
-		private LibraryInfo(org.openlca.core.library.LibraryInfo info, List<String> linkedIn) {
-			this(info.name(), info.description(), info.isRegionalized(), linkedIn);
+		private LibraryInfo(org.openlca.core.library.LibraryInfo info, List<String> linkedIn,
+				List<String> accessTypes) {
+			this(info.name(), info.description(), info.isRegionalized(), linkedIn, accessTypes);
 		}
 
 	}
